+++ ComparisonOperator.pde
@@ -1,5 +1,5 @@
 abstract class ComparisonOperator extends PrimeOperator{
   void initialize(){
-    addInPork(DataCategory.FLOAT); addInPork(DataCategory.FLOAT); addOutPork(DataCategory.BOOL).setTargetFlow(new Flow(DataCategory.BOOL));
+    addInPork(DataCategory.NUMERIC); addInPork(DataCategory.NUMERIC); addOutPork(DataCategory.BOOL).setTargetFlow(new Flow(DataCategory.BOOL));
   }
 }
