+++ UnaryOperator.pde
@@ -6,7 +6,7 @@ public abstract class UnaryOp extends PrimeOperator implements UnaryOperator<Flo
   }
   
   void initialize(){
-    addInPork(DataCategory.FLOAT); addOutPork(DataCategory.FLOAT);
+    addInPork(DataCategory.NUMERIC); addOutPork(DataCategory.NUMERIC);
   }
 
   public abstract float apply(float value);
